Faster memory free iterator

+++ core/common-core/src/main/java/eu/leads/processor/infinispan/IntermediateKeyIndex.java
@@ -10,7 +10,7 @@ import java.util.Set;
  * Created by vagvaz on 16/07/15.
  */
 public class IntermediateKeyIndex {
-    private Map<String,Long> keysCache;
+    private Map<String,Integer> keysCache;
     private Map<String,Object> dataCache;
 
     public IntermediateKeyIndex(Cache keysCache, Cache dataCache) {
@@ -19,9 +19,9 @@ public class IntermediateKeyIndex {
     }
 
     public void put(String key, Object value){
-        Long count = keysCache.get(key);
+        Integer count = keysCache.get(key);
         if(count == null){
-            count = 0L;
+            count = 0;
         }
         else{
             count++;
@@ -30,11 +30,11 @@ public class IntermediateKeyIndex {
         keysCache.put(key,count);
     }
 
-    public Set<Map.Entry<String,Long>> getKeysIterator(){
+    public Set<Map.Entry<String,Integer>> getKeysIterator(){
         return keysCache.entrySet();
     }
 
-    public Iterator<Object> getKeyIterator(String key, Long counter){
+    public Iterator<Object> getKeyIterator(String key, Integer counter){
         return new LocalIndexKeyIterator(key,counter,dataCache);
     }
